requirements.txt
httpx
pytest-asyncio
fastapi
uvicorn
